refactor(admin): address review — shared FieldRow, honest partial workshop window
Promote FieldRow out of the settings route module into
src/components/admin/FieldRow.tsx (settingsLayout re-exports it) so the
client tiers page no longer imports from an app route. Normalize null
registration bounds at the workshops page boundary. Derive a distinct
'Partially configured' status (with an enforcement-accurate note) when only
one bound is set, mirroring the independent-bound checks in the signup path,
and pin the clock in the new stories.
